chore: rename applicant script.js to applicant.js
https://github.com/POWERLANE-TEAM/emplo-hrms/pull/173#discussion_r1883974049
