+++ resources/js/applicant/applicant.js
