+++ resources/js/applicant/dashboard.js
@@ -1,7 +1,7 @@
 
 import "../../css/applicant/dashboard.css";
 import '../script.js';
-import './script.js';
+import './applicant.js';
 import initLucideIcons from '../icons/lucide.js';
 import addGlobalScrollListener, { documentScrollPosY } from 'global-scroll-script';
 import addGlobalListener, { GlobalListener } from 'globalListener-script';
